Completed manage spots, update feature

+++ authenticate-me/backend/db/migrations/20230124191206-create-spot.js
@@ -58,7 +58,7 @@ module.exports = {
         allowNull: false
       },
       price: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(6,2),
         allowNull: false
       },
       createdAt: {
