+++ authenticate-me/backend/db/models/spot.js
@@ -106,7 +106,7 @@ module.exports = (sequelize, DataTypes) => {
       }
     },
     price: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.DECIMAL(6,2),
       allowNull: false,
       validate: {
         min: 0
